+++ src/App.jsx
@@ -1,6 +1,6 @@
 import About from "./components/About/About";
 import Contact from "./components/Contact";
-import Projects from "./components/Projects";
+import Projects from "./components/Projects/Projects";
 
 const App = () => {
   return (
